refactor: expand completion traces
