+++ packages/base-types/src/trace/completion/completion-end-trace.ts
@@ -1,0 +1,7 @@
+import { BaseTraceFrame, TraceType } from '@base-types/node/utils';
+
+export interface CompletionEndTrace extends BaseTraceFrame {
+  type: TraceType.COMPLETION_END;
+}
+
+export const isCompletionEndTrace = (trace: BaseTraceFrame): trace is CompletionEndTrace => trace.type === TraceType.COMPLETION_END;
